+++ config/app.php
@@ -269,6 +269,26 @@ return [
 
 	/*
 	|--------------------------------------------------------------------------
+	| URLs
+	|--------------------------------------------------------------------------
+	|
+	| URLs that are used throughout the interface.
+	|
+	*/
+	'urls' => [
+		'home' => 'https://meta.miraheze.org/wiki/Special:MyLanguage/Miraheze_Meta',
+		'mediawiki' => [
+			'api' => 'https://login.miraheze.org/w/api.php',
+			'centralauth' => 'https://meta.miraheze.org/wiki/Special:CentralAuth',
+		],
+		'policies' => [
+			'privacy' => 'https://meta.miraheze.org/wiki/Special:MyLanguage/Privacy_Policy',
+			'tou' => 'https://meta.miraheze.org/wiki/Special:MyLanguage/Terms_of_Use',
+		],
+	],
+
+	/*
+	|--------------------------------------------------------------------------
 	| At Risk Emails
 	|--------------------------------------------------------------------------
 	|