+++ resources/views/report/view.blade.php
@@ -87,7 +87,7 @@
 						@endcan
 						<div class="d-flex flex-column gap-1">
 							<span class="fw-semibold text-dark">{{ __( 'centralauth' ) }}:</span>
-							<a class="btn btn-outline-primary btn-sm text-truncate" href="https://meta.miraheze.org/wiki/Special:CentralAuth/{{ $report->user->username }}" style="max-width: 250px;">
+							<a class="btn btn-outline-primary btn-sm text-truncate" href="{{ config( 'app.urls.mediawiki.centralauth' ) }}/{{ $report->user->username }}" style="max-width: 250px;">
 								{{ $report->user->username }}
 							</a>
 						</div>